Do not longer support time_t aka int ast TimeStamp
The mtime at logfile is a 100ns value used to compare the ftime values
with same resolution

Conflicts:
	.gitignore
	configure.py
	src/build.cc
	src/build_log_perftest.cc
	src/build_test.cc
	src/canon_perftest.cc
	src/hash_map.h
	src/metrics.cc
	src/parser_perftest.cc

+++ src/canon_perftest.cc
@@ -1,4 +1,3 @@
-#include <string.h> // for strlen, strcopy
 #include "util.h"
 
 const char kPath[] =
@@ -25,9 +24,9 @@ int main() {
 
   int min = times[0];
   int max = times[0];
-  float total = 0;
+  float total = 0.0;
   for (size_t i = 0; i < times.size(); ++i) {
-    total += times[i];
+    total += static_cast<float>(times[i]);
     if (times[i] < min)
       min = times[i];
     else if (times[i] > max)
@@ -35,5 +34,5 @@ int main() {
   }
 
   printf("min %dms  max %dms  avg %.1fms\n",
-         min, max, total / times.size());
+         min, max, total / static_cast<float>(times.size()));
 }
